Drop unnecessary and wrong group detection on host
Don't call get_group_for_sudo() on the host during create(). That runs
on the host, and thus will check which sudo group exists on the host.
But that is entirely irrelevant for sudo inside the container, and it
breaks when trying to create a Debian or Ubuntu based toolbox on a
Fedora host (or vice versa). This also causes problem on CoreOS[0][1]

Also, there is no point in running the `podman create` command with an
extra sudo group, normal user privileges are just fine.

init_container() will call get_group_for_sudo() inside the container and
initialize the groups correctly there.

https://github.com/containers/toolbox/pull/401

[0] https://github.com/containers/toolbox/issues/423
[1] https://github.com/coreos/fedora-coreos-tracker/issues/458